Refactor onboarding UI structure
